processing/requirements.txt
kafka-python
scikit-learn
nltk
numpy